This code differentiates the input signal.
Analog or digital input signal is provided to MSP430F5529. It
differentiates the input signal.
First, the ADC operation is performed.
Samples are stored in DMA registers.
Samples are converted into voltages and stored in an array.
Differentiator function is performed on this array and results are
stored in other array.